Modificacione en el campo referencia del objeto account.bank.statement, ahora se guarda un char con el numero de recibo u orden de pago

+++ l10n_ar_bank_statement/bank_statement.py
@@ -59,7 +59,12 @@ class account_bank_statement_line(osv.osv):
             st_lines = context.get('st_lines',False)
             journal_type = context.get('journal_type',False)
             
-        if st_lines:
+        if st_lines and st_lines in 'remove':
+            for id in ids:
+                sql = 'delete from account_bank_statement_line where id = ' + str(id)
+                cr.execute(sql)
+            return {}
+        elif st_lines and st_lines in 'delete':
             for t in self.browse(cr, uid, ids, context=context):
                 if t.state not in 'draft' or t.creation_type in 'system':
                     raise osv.except_osv(_('Invalid action !'), _('Cannot delete Account Bank Statement Line(s) which are not draft state o creation type is system !'))
@@ -78,12 +83,16 @@ class account_bank_statement_line(osv.osv):
                 else:
                     t = self.browse(cr, uid, ids, context=context)
 
-                    if t.state not in ('draft'):
-                        raise osv.except_osv(_('Invalid action !'), _('Cannot delete Account Bank Statement Line(s) which are not draft state !'))
+                    if t.state not in ('draft') or t.creation_type in 'manual':
+                        raise osv.except_osv(_('Invalid action !'), _('Cannot delete Account Bank Statement Line(s) which are not draft state or creation type manual!'))
                     sql = 'delete from account_bank_statement_line where id = ' + str(t['id'])
                     cr.execute(sql)
                     return {}
             else:
+                #~ t = self.browse(cr, uid, ids, context=context)
+#~ 
+                #~ if t.state not in ('draft') or t.creation_type in 'manual':
+                    #~ raise osv.except_osv(_('Invalid action !'), _('Cannot delete Account Bank/Cash Statement Line(s) which are not draft state or creation type manual!'))
                 sql = 'delete from account_bank_statement_line where id = ' + str(ids)
                 cr.execute(sql)
                 return {}
